+++ tests/conftest.py
@@ -1,9 +1,3 @@
-# Adding pytest base dir to Python system path.
-# This is required in order to import from common package including pytest_plugins within this file.
-import site
-from os.path import dirname, abspath
-site.addsitedir(dirname(abspath(__file__)))
-
 import sys
 import os
 import glob
@@ -92,24 +86,16 @@ def pytest_addoption(parser):
     # test_vrf options
     parser.addoption("--vrf_capacity", action="store", default=None, type=int, help="vrf capacity of dut (4-1000)")
     parser.addoption("--vrf_test_count", action="store", default=None, type=int, help="number of vrf to be tested (1-997)")
-
     ############################
     # test_techsupport options #
     ############################
+
     parser.addoption("--loop_num", action="store", default=10, type=int,
                     help="Change default loop range for show techsupport command")
     parser.addoption("--loop_delay", action="store", default=10, type=int,
                     help="Change default loops delay")
     parser.addoption("--logs_since", action="store", type=int,
                     help="number of minutes for show techsupport command")
-
-    ############################
-    #   sanity_check options   #
-    ############################
-    parser.addoption("--skip_sanity", action="store_true", default=False,
-                     help="Skip sanity check")
-    parser.addoption("--allow_recover", action="store_true", default=False,
-                     help="Allow recovery attempt in sanity check in case of failure")
 
 
 @pytest.fixture(scope="session", autouse=True)
