+++ tests/platform/conftest.py
@@ -1,6 +1,6 @@
 import pytest
 from common.fixtures.advanced_reboot import get_advanced_reboot
-from .args.advanced_reboot_args import add_advanced_reboot_args
+from platform_args.advanced_reboot_args import add_advanced_reboot_args
 
 @pytest.fixture(autouse=True, scope="module")
 def skip_on_simx(duthost):
