+++ tests/platform/mellanox/test_thermal_control.py
@@ -6,7 +6,7 @@ import time
 from common.mellanox_data import SWITCH_MODELS
 from common.plugins.loganalyzer.loganalyzer import LogAnalyzer
 from common.utilities import wait_until
-from ..thermal_control_test_helper import *
+from thermal_control_test_helper import *
 from mellanox_thermal_control_test_helper import MockerHelper, AbnormalFanMocker
 
 THERMAL_CONTROL_TEST_WAIT_TIME = 65
